requirements.txt
fastapi
uvicorn
requests
qrcode
python-stdnum
textblob
dnspython
pydantic[email]
python-multipart
